Load sample time collected from example data
Add the Time Collected column to the example sample table data

Load the time collected as a sample attribute

Time collected data for some additional experiments (not included in the
PR) are floats (_e.g._ `1.5` mins past infusion), so the
`Sample.time_collected` field was changed to be a float instead of an
integer

+++ DataRepo/models.py
@@ -175,7 +175,7 @@ class Sample(models.Model):
     """
     MINIMUM_VALID_TIME_COLLECTED = -1440
     MAXIMUM_VALID_TIME_COLLECTED = 10080
-    time_collected = models.SmallIntegerField(
+    time_collected = models.FloatField(
         null=True,
         blank=True,
         validators=[
